requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cors-headers==3.3.0
djangorestframework==3.11.0
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
